selected flow state in not reset on route change
